add Header.js and move function
