+++ js-structure/js-structure_quiz-app/components/Header/Header.js
@@ -1,0 +1,9 @@
+export default function Header() {
+    const header = document.createElement("header");
+    header.classList.add("header");
+    header.innerHTML = /* html */ `
+        <h1 class="header__title">Quiz-App</h1>
+      `;
+  
+    return header;
+  }
