+++ js-structure/js-structure_quiz-app/index.js
@@ -1,3 +1,5 @@
+import Header from "./components/Header/Header.js";
+
 // Store the cards in a global state
 const cards = [
   {
@@ -23,15 +25,7 @@ const cards = [
   },
 ];
 
-function Header() {
-  const header = document.createElement("header");
-  header.classList.add("header");
-  header.innerHTML = /* html */ `
-	  <h1 class="header__title">Quiz-App</h1>
-	`;
 
-  return header;
-}
 
 function handleFormSubmit(event) {
   event.preventDefault();
